chore: enable strict compiler warnings and fix all diagnostics
- Add -Wconversion, -Wsign-conversion, -Wshadow, -Wnon-virtual-dtor,
  -Woverloaded-virtual, -Wimplicit-fallthrough, -Wnull-dereference,
  -Wold-style-cast, -Wformat=2, -Wcast-align, -Wdouble-promotion,
  -Wundef, -Wextra-semi, -Wduplicated-cond, -Wduplicated-branches,
  -Wlogical-op, -Wuseless-cast, -Wsuggest-override to GCC/Clang builds
- Fix all -Wconversion warnings (qsizetype -> int) with static_cast<int>()
- Fix -Wuseless-cast warnings in BeWavedDolphin and Led

+++ App/Element/GraphicElements/InputRotary.cpp
@@ -294,7 +294,7 @@ void InputRotary::setSkin(const bool defaultSkin, const QString &fileName)
 
 int InputRotary::outputSize() const
 {
-    return outputs().size();
+    return static_cast<int>(outputs().size());
 }
 
 int InputRotary::outputValue() const
